feat: Requisito 13
endpoint PUT /cars/id

+++ src/models/MongoModel.ts
@@ -1,4 +1,4 @@
-import { isValidObjectId, Model } from 'mongoose';
+import { isValidObjectId, Model, UpdateQuery } from 'mongoose';
 import { ErrorTypes } from '../errors/catalog';
 import { IModel } from '../interfaces/IModel';
 
@@ -23,10 +23,18 @@ abstract class MongoModel<T> implements IModel<T> {
   }
 
   public async update(_id: string, obj: T): Promise<T | null> {
-    return this._model.findByIdAndUpdate({ _id }, { obj });
+    if (!isValidObjectId(_id)) throw new Error(ErrorTypes.InvalidMongoId);
+
+    return this._model.findByIdAndUpdate(
+      { _id }, 
+      { ...obj } as UpdateQuery<T>,
+      { new: true },
+    );
   }
 
   public async delete(_id: string): Promise<T | null> {
+    if (!isValidObjectId(_id)) throw new Error(ErrorTypes.InvalidMongoId);
+
     return this._model.findByIdAndDelete({ _id });
   }
 }
